Introduce new abstract parent class for definitions page
